rework sidebar
